Support compiling arithemtic operations for the `i8` type. Fix a bug in `i64` division, this issue occurs due to zeroing out the rdx register even if the fist operand was negative, it should be sign extended instead.

+++ src/runtime.c
@@ -1,11 +1,12 @@
 // Buraq Runtime
 
 #include <stdio.h>
+#include <stdint.h>
 
 // this is the entry function to buraq
-extern int boot() asm("boot");
+extern int64_t boot() asm("boot");
 
 int main() {
-    printf("%d\n", boot());
+    printf("%lld\n", boot());
     return 0;
 }
